[network\ndis\ndisprot] Virtual Miniport: Fix UINT8 comparison with ULONG (#754)
fix UINT8 comparison with ULONG

+++ network/ndis/netvmini/6x/rssv2.c
@@ -210,7 +210,7 @@ Return Value:
 
 --*/
 {
-    UINT8 index;
+    ULONG index;
     PNDIS_RSS_PROCESSOR processor;
 
     for (index = 0;
